feat(template.html): Apply changes from upstream (styles.html partial)
As mentioned in the previous commit, we now deal with the changes that came with Pandoc 2.8 (released 2019-11-22).

Basically the CSS code was decoupled from the HTML template and put into the `styles.html` "partial" (partials = _subtemplates stored in different files_, as explained in the [manual](https://pandoc.org/MANUAL.html#partials).

The partial needs to be in the same directory as the template file, so there I created it and filled it with the CSS code extracted from the `template.html` file.

This makes it almost -- but not exactly similar to the [initial contents](https://github.com/jgm/pandoc-templates/commit/6e96149c63f4fe357f09976ca3da1136877b5198) of the `styles.html` file in the `pandoc-templates` repository: two lines are missing, that I'll add in the next commit.

Consequently, `styles.html` is added to the dependencies of the `docs/index.html` rule.