feat(template.html): Output MathJax script using command-line
This commit replaces the "hardwired" inclusion of MathJax in the `template.html` with the `--mathjax` option in the Pandoc evocation, with does the same.

This can be verified by doing
```sh
echo '$a

|pandoc -f markdown -t html5 -s --mathjax
```
and noticing that the output contains
```html
<script src="https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-chtml-full.js" type="text/javascript"></script>
```

As the `--mathjax` already belongs to the `pandoc_args`, the `Makefile` stays unchanged.