feat(template.html): Add a scripts.html partial
Finally, I extracted the extra `<script>` markups into a new partial file `scripts.html`.
Similarly to with `styles.html`, the partial is added to the dependencies of the `docs/index.html` rule.

This makes the `template.html` very close to its upstream counterpart, and very easy to compare and keep on par with further changes from upstream.